+++ stubs/app/Providers/AppServiceProvider.stub.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\Providers;
+
+use App\Models\Team;
+use App\Models\User;
+use App\Policies\TeamPolicy;
+use BrilliantPortal\Framework\Policies\UserPolicy;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\ServiceProvider;
+
+class AppServiceProvider extends ServiceProvider
+{
+    /**
+     * Register any application services.
+     */
+    public function register(): void
+    {
+        //
+    }
+
+    /**
+     * Bootstrap any application services.
+     */
+    public function boot(): void
+    {
+        //
+
+        Gate::policy(User::class, UserPolicy::class);
+        Gate::policy(Team::class, TeamPolicy::class);
+    }
+}